Move `FieldPath` expressions to `expression` package
