+++ java/src/main/kotlin/io/spine/validation/java/expression/FieldPaths.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java.violation
+package io.spine.validation.java.expression
 
 import io.spine.base.FieldPath
 import io.spine.protodata.ast.FieldName
@@ -32,7 +32,6 @@ import io.spine.protodata.java.Expression
 import io.spine.protodata.java.StringLiteral
 import io.spine.protodata.java.call
 import io.spine.protodata.java.toBuilder
-import io.spine.validation.java.expression.FieldPathsClass
 
 /**
  * Returns an expression that yields this [FieldPath] as a string using
@@ -46,9 +45,9 @@ internal fun Expression<FieldPath>.joinToString(): Expression<String> =
 
 /**
  * Returns an expression that yields a new instance of [FieldPath] by appending
- * the provided [field] name to the provided [parent] path.
+ * the provided [field] name to this parental [FieldPath] expression.
  */
-internal fun fieldPath(parent: Expression<FieldPath>, field: FieldName): Expression<FieldPath> =
-    parent.toBuilder()
+internal fun Expression<FieldPath>.resolve(field: FieldName): Expression<FieldPath> =
+    toBuilder()
         .chainAdd("field_name", StringLiteral(field.value))
         .chainBuild()
