+++ java/src/main/kotlin/io/spine/validation/java/generate/option/DistinctFieldGenerator.kt
@@ -54,8 +54,8 @@ import io.spine.validation.java.expression.MultiSetEntryClass
 import io.spine.validation.java.generate.ValidationCodeInjector.MessageScope.message
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.violation.constraintViolation
-import io.spine.validation.java.violation.fieldPath
-import io.spine.validation.java.violation.joinToString
+import io.spine.validation.java.expression.joinToString
+import io.spine.validation.java.expression.resolve
 import io.spine.validation.java.expression.stringValueOf
 import io.spine.validation.java.generate.FieldOptionGenerator
 import io.spine.validation.java.violation.templateString
@@ -82,7 +82,7 @@ internal class DistinctFieldGenerator(private val view: DistinctField) : FieldOp
             """
             if (!$collection.isEmpty() && $collection.size() != $set.size()) {
                 var duplicates = ${extractDuplicates(collection)};
-                var fieldPath = ${fieldPath(parentPath, field.name)};
+                var fieldPath = ${parentPath.resolve(field.name)};
                 var violation = ${violation(ReadVar("fieldPath"), ReadVar("duplicates"))};
                 violations.add(violation);
             }
